requirements.txt
discord.py==2.3.2
gTTS==2.5.1
translate==3.6.1
PyNaCl==1.5.0
python-dotenv==1.0.1
langcodes[data]